Splittet previous function into two - added formBuilder
